Refactore code
