+++ transactions/serializers.py
@@ -5,7 +5,7 @@ from transactions.models import ClientAccount, Deposit, Transaction, Savings
 from rest_framework.exceptions import ValidationError
 from property.models import Property
 from property.serializers import PropertySerializer
-from authentication.models import User
+from authentication.models import User, CardInfo
 
 
 class ClientAccountSerializer(serializers.ModelSerializer):
@@ -233,3 +233,13 @@ class DepositSerializer(serializers.ModelSerializer):
             "saving_account",
             "transaction"
         ]
+
+
+class CardInfoSerializer(serializers.ModelSerializer):
+    """
+    Handle serializing and deserializing  of card info
+    """
+
+    class Meta:
+        model = CardInfo
+        fields = '__all__'
