+++ transactions/urls.py
@@ -9,7 +9,9 @@ from transactions.views import (
     RetreiveTransactionsAPIView,
     foreign_card_validation_response,
     tokenized_card_payment,
-    RetrieveDepositsApiView
+    RetrieveDepositsApiView,
+    SavedCardsListView,
+    DeleteSavedCardView
 )
 
 app_name = 'transactions'
@@ -31,5 +33,10 @@ urlpatterns = [
     path(
          'tokenized-card/<int:saved_card_id>',
          tokenized_card_payment, name='tokenized_card'),
-    path('my-deposit/', RetrieveDepositsApiView.as_view(), name='my_deposit')
+    path('my-deposit/', RetrieveDepositsApiView.as_view(), name='my_deposit'),
+        path('saved-cards/',
+         SavedCardsListView.as_view(), name='saved-cards'),
+    path(
+         'saved-card/<int:id>',
+         DeleteSavedCardView.as_view(), name='saved-card'),
 ]
